refactor: transfer the multistep modal code from the TestHomePage to the current HomePage
